+++ DirectDM/Kernel/init.m
@@ -3,6 +3,7 @@
 Get["DirectDM`usage`"];
 Get["DirectDM`functions`"];
 Get["DirectDM`inputs`"];
+Get["DirectDM`formfactors`"];
 
 
 (* Bases *)